+++ example_process_streams.py
@@ -44,7 +44,7 @@ except ImportError:
 
 class BinanceWebSocketApiProcessStreams(object):
     @staticmethod
-    def process_stream_data(received_stream_data_json, exchange="binance.com", stream_buffer_name="False"):
+    def process_stream_data(received_stream_data_json, stream_buffer_name="False"):
         #
         #  START HERE!
         #
@@ -58,6 +58,8 @@ class BinanceWebSocketApiProcessStreams(object):
         # to see the difference.
         # Github: https://github.com/LUCIT-Systems-and-Development/unicorn-fy
         # PyPI: https://pypi.org/project/unicorn-fy/
+        exchange = "binance.com"
+
         if exchange == "binance.com" or exchange == "binance.com-testnet":
             unicorn_fied_stream_data = UnicornFy.binance_com_websocket(received_stream_data_json)
         elif exchange == "binance.com-futures" or exchange == "binance.com-futures-testnet":