+++ unicorn_binance_websocket_api/sockets.py
@@ -217,13 +217,13 @@ class BinanceWebSocketApiSocket(object):
                         logger.error("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
                                      str(self.channels) + ", " + str(self.markets) + ") - KeyError (possibly within the"
                                      "callback function) - error_msg: " + str(error_msg))
-                    except Exception as error_msg:
-                        logger.error("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
-                                     str(self.channels) + ", " + str(self.markets) + ") - Exception General Exception "
-                                     " (possibly within the callback function) - error_msg: " + str(error_msg))
-                        self.manager.stream_is_crashing(self.stream_id, str(error_msg))
-                        self.manager.set_restart_request(self.stream_id)
-                        sys.exit(1)
+#                    except Exception as error_msg:
+#                        logger.error("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
+#                                     str(self.channels) + ", " + str(self.markets) + ") - Exception General Exception "
+#                                     " (possibly within the callback function) - error_msg: " + str(error_msg))
+#                        self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+#                        self.manager.set_restart_request(self.stream_id)
+#                        sys.exit(1)
         except asyncio.TimeoutError as error_msg:
             # Catching https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/issues/221
             self.manager.stream_is_crashing(self.stream_id, error_msg)