+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Discussions about unicorn-binance-websocket-api releases!](https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/discussions/categories/releases)
 
 ## 1.40.2.dev (development stage/unreleased/unstable)
+### Fixed
+- Catch KeyError before the general exception in sockets to give better feedback within the callback.
 
 ## 1.40.2
 ### Changed